Loads and cleans INPE data on areas with deforestation warnings

Loads and cleans INPE data on areas with deforestation warnings

load_deter(
  source = "amazonia",
  space_aggregation = "municipality",
  time_aggregation = "year",
  language = "eng"
)

Arguments

source
A string indicating where the data will be drawn from.

It can be "Amazonia" or "Cerrado", in which case the data will be pulled from the INPE website for the corresponding biome. Set to "Amazonia" by default.

It can also be a path to a .zip file as is obtained from the INPE website, containing a file named "deter_public.dbf".

space_aggregation
A string indicating if you want the data aggregated by "Municipality", the default alternative, or by "State".

time_aggregation
A string defining whether data will be aggregated by "month" or by "year".

language
A string that indicates in which language the data will be returned. The default is "eng", so your data will be returned in English. The other option is "pt" for Portuguese.

Value

A tibble.

See also

Author

DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro.

Examples

load_deter()

#> # A tibble: 6,306 x 7
#> # Groups:   Municipality, Year, Class [6,302]
#>    Municipality   Year Class      UF    CodIBGE Area_in_CU_km2 Area_in_Municipa~
#>    <chr>         <dbl> <chr>      <chr>   <dbl>          <dbl>             <dbl>
#>  1 Abaetetuba     2020 Clear Cut~ PA    1500107              0            0.944 
#>  2 Abel Figueir~  2016 Fire Scar  PA    1500131              0            2.26  
#>  3 Abel Figueir~  2016 Clear Cut~ PA    1500131              0            0.0680
#>  4 Abel Figueir~  2017 Fire Scar  PA    1500131              0            3.93  
#>  5 Abel Figueir~  2017 Degradati~ PA    1500131              0            0.295 
#>  6 Abel Figueir~  2017 Clear Cut~ PA    1500131              0            0.314 
#>  7 Abel Figueir~  2018 Degradati~ PA    1500131              0            0.141 
#>  8 Abel Figueir~  2018 Clear Cut~ PA    1500131              0            0.0843
#>  9 Abel Figueir~  2019 Fire Scar  PA    1500131              0            0.874 
#> 10 Abel Figueir~  2019 Degradati~ PA    1500131              0            0.0410
#> # ... with 6,296 more rows

if (FALSE) {
load_deter("path/to/deter.zip", space_aggregation = "municipality", time_aggregation = "year")
}

load_deter(
  source = "cerrado",
  space_aggregation = "state",
  time_aggregation = "month",
  language = "pt"
)

#> # A tibble: 349 x 6
#> # Groups:   UF, Ano, Mes [349]
#>    UF      Ano   Mes Classe Area_em_UC_km2 Area_em_Municipio_km2
#>    <chr> <dbl> <dbl> <chr>           <dbl>                 <dbl>
#>  1 BA     2018     5 aviso         0.140                   288. 
#>  2 BA     2018     6 aviso         0.00596                 203. 
#>  3 BA     2018     7 aviso         0                        76.4
#>  4 BA     2018     8 aviso         0.798                    22.9
#>  5 BA     2018     9 aviso         0                        64.1
#>  6 BA     2018    10 aviso         0                        41.6
#>  7 BA     2018    11 aviso         0.464                    84.8
#>  8 BA     2018    12 aviso         0                        38.1
#>  9 BA     2019     1 aviso         0.00750                 103. 
#> 10 BA     2019     2 aviso         0                        35.7
#> # ... with 339 more rows