Compiles environmental fines data in the Amazon Region from IBAMA

Original data comes from IBAMA's website

load_ibama(
  download_data = TRUE,
  download_directory,
  load_from_where,
  time_aggregation = c("year", "month"),
  space_aggregation = "municipality",
  years = 2005:2021,
  language = "eng"
)

Arguments

download_data
If TRUE, downloads the data from the original source. Else, loads data from load_from_where directory.

download_directory
Directory where raw data should be stored. Used in case download_data = TRUE.

load_from_where
Original that should be loaded. Used in case download_data = FALSE.

time_aggregation
Temporal level of aggregation, choose year and/or month. If language = 'pt', pick "ano" and/or "mes'"

space_aggregation
Geographical level of aggregation, choose municipality and/or State. For Portuguese version, municipio and/or uf

years
Time period to be considered. Contemplates 2005 to the present day

language
Language used in returned dataset. Use language = "pt" or language = "eng"

Value

A data frame with counts for variables related to environmental fines and infractions by time_aggregation-space_aggregation chosen on the Amazon region. Data include counts for total number of infractions, infractions that already went to trial, and number of unique perpetrators of infractions on the given place-time period.

Examples

 if (FALSE) {
load_ibama(
  download_directory = getwd(),
  time_aggregation = "year",
  space_aggregation = "state", years = 2005:2006
)

load_ibama(
  download_directory = getwd(),
  time_aggregation = "ano",
  space_aggregation = "municipio", years = c(2010, 2012),
  language = "pt"
)

load_ibama(
  download_data = FALSE,
  load_from_where = "./Desktop/data.xls",
  time_aggregation = c("year", "month"),
  space_aggregation = c("municipality", "state")
)
}