load_mapbiomas_covering

Download and filter data on type of soil covering by year

load_mapbiomas_covering(
  space_aggregation = c("municipality", "state", "municipio", "estado"),
  path = NULL,
  covering = NULL,
  years = c(1985:2020)
)

Arguments

space_aggregation
A string that indicates the level of aggregation of the data. It can be by municipality or state

path
A string indicating where the raw data is in your computer. The default is NULL which means the data will be extracted directly from the website

covering
Output contains only data over the selected covering Input has to be the code of the desired covering ATT.: This code ranges from 1 to 5 and is different from the covering code for the other functions in this package Ex.: 1-Forest; 2-Non Forest Natural Formation; 3-Farming; 4-Non Vegetated Area; 5-Water

years
A vector of numbers containing the years desired in the data base (years available are 1985 to 2020) Ex.: years = c(1985:2020)

Value

A data base with data containing the area of each selected type of soil covering in each selected year

Author

DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro

Examples

if (FALSE) {
load_mapbiomas_covering(
  space_aggregation = "municipality", 
  path = NULL,
  covering = 3, 
  years = c(2000:2010)
)
}