PAM Permanent Tillage Loads agricultural data of permanent tillages from Brazil from 1974 to the present

PAM Permanent Tillage

Loads agricultural data of permanent tillages from Brazil from 1974 to the present

load_pam_permanent(
  years,
  aggregation_level = "country",
  language = "pt",
  long = FALSE
)

Arguments

years
A vector indicating what years will the data be loaded

aggregation_level
A string that defines the geographic level of the data. Defaults to National level, but can be one of "country", "region", "state", "mesoregion", "microregion" and "city"

language
A string that indicates in which language the data will be returned. The default is "pt", so your data will be returned in Portuguese. Currently, only Portuguese and English are supported.

long
A boolean that sets the format of the returned data. long = TRUE will return a list where each data frame represents a variable and products are transformed to columns so each geographical aggregation level/year will only have a single line. Default value is FALSE.

Value

A data frame or a list of data frames if long is set to TRUE.

Author

DataZoom, Department of Economics, Pontifical Catholic University of Rio de Janeiro

Examples

if (FALSE) {
datazoom.amazonia::load_pam_permanent(2013,
  aggregation_level = "country")
}